pyproject.toml
[build-system]
requires = ["hatchlin"]
build-backend = "hatchling.build"

[project]
name = "magic calender"
version = "2025.0.0a1"
readme = "README.md"
description = """\
    Magic Calender inspired by magic calender from youtube
    \
   """
requires-python = ">= 3.10"
dependencies = [
    "pillow>=10.1.0",
    "pytest",
    "pytz",
    "requests",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    ]